refactor: move named test modules into sibling files
Extract the `#[cfg(test)] mod <name>_tests { ... }` blocks (which the
first sweep skipped because they are not named `tests`) into sibling
files declared `#[cfg(test)] mod <name>_tests;`. A named module relocated
to its own same-named file keeps its parent, so `super::` paths are
unchanged and the bodies move verbatim.

Covers compress_bound / hc / tag_mask / neon_tag_mask / ldm_helper /
extend_with_repcode / bt_pair_index_wrap / stream_abs_headroom / storage
/ zerocopy_robustness / burst_gate / frame_inspection / inline_helper /
portable_helper tests across the encoding and decoding trees. Suite
unchanged at 824 cases; x86_64 build cross-checked for the arch-gated
tag-mask tests.
